requirements.txt
pandas>=2.0.0
selenium>=4.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
